using hardcoded values for initial testing with front end while we do mongodb setup
