+++ closetr-api/models/clothes.js
@@ -4,7 +4,7 @@ const clothes = new mongoose.Schema({
   name: {type: String, default:'', trim:true},
   cost: {type: Number, default:0},
   worn_count: {type: Number, default:0},
-  clothing_category: {type: String, default:'', trim:true}
+  category: {type: String, default:'', trim:true}
 })
 
 module.exports = mongoose.model('clothes', clothes)
